Adjust patch paths to the new upstream Git repository layout
The new upstream "msys-runtime" repository only contains the equivalent to
what was previously available inside the /msys/rt/src/ directory of our
CVS mirror, see

    http://repo.or.cz/w/msys.git/tree/b4ca4aaba1c45ee13cb52045aaf3d726f73e5e1d:/msys/rt/src

So as there is only the runtime and no applications we need to drop the
bash patch from here. If not contributed upstream in between, the bash
patch should be added again as part of a new /src/bash/ directory.

Signed-off-by: Sebastian Schuberth <sschuberth@gmail.com>